refactor: project folder structure and organisations
