+++ lib/models/models.dart
@@ -1,0 +1,4 @@
+export 'contacts_model.dart';
+export 'appointments_model.dart';
+export 'notes_model.dart';
+export 'task_model.dart';
