+++ lib/ui/contacts/contacts.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'contacts_entry.dart';
 import 'contacts_list.dart';
 import 'package:scoped_model/scoped_model.dart';
-import 'contacts_model.dart' show contactsModel, ContactsModel;
+import '../../models/models.dart' show contactsModel, ContactsModel;
 
 class Contacts extends StatelessWidget {
   const Contacts({Key? key}) : super(key: key);
