+++ lib/ui/main_screens.dart
@@ -1,0 +1,4 @@
+export 'appointments/appointments.dart';
+export 'contacts/contacts.dart';
+export 'notes/notes.dart';
+export 'tasks/tasks.dart';
