+++ lib/ui/tasks/tasks.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
+import '../../models/models.dart' show TaskModel, tasksModel;
 import 'tasks_entry.dart';
 import 'tasks_list.dart';
 import 'package:scoped_model/scoped_model.dart';
-import 'task_model.dart' show tasksModel, TaskModel;
 
 class Tasks extends StatelessWidget {
   const Tasks({Key? key}) : super(key: key);
